+++ examples/sse/src/main.rs
@@ -4,8 +4,7 @@
 //! cargo run -p sse
 //! ```
 
-use volga::{App, error::Error, http::sse::Message, sse};
-use futures_util::stream::{repeat_with};
+use volga::{App, http::sse, sse};
 use std::time::Duration;
 use serde::Serialize;
 use tokio_stream::StreamExt;
@@ -20,18 +19,15 @@ async fn main() -> std::io::Result<()> {
     let mut app = App::new();
 
     app.map_get("/events", || async {
-        let repeater = || {
-            let resp = SseResponse { data: "Hello, world!".into() };
-            Message::new()
-                .comment("A greeting message")
-                .event("greet")
-                .json(resp)
-                .retry(Duration::from_secs(3))
-        };
-        let stream = repeat_with(repeater)
-            .map(Ok::<_, Error>)
+        let resp = SseResponse { data: "Hello, world!".into() };
+        let stream = sse::Message::new()
+            .comment("A greeting message")
+            .event("greet")
+            .json(resp)
+            .retry(Duration::from_secs(3))
+            .repeat()
             .throttle(Duration::from_secs(1));
-
+        
         sse!(stream)
     });